+++ src/emit/index.ts
@@ -33870,7 +33870,6 @@ class Emitter {
         const bodyIsContinue = ts.isContinueStatement(bodyAction) && !bodyAction.label;
         const bodyIsBreak = ts.isBreakStatement(bodyAction) && !bodyAction.label;
         if (!bodyIsContinue && !bodyIsBreak) return false;
-        if (loopInitializer && !bodyIsBreak && ts.isVariableStatement(loopInitializer)) return false;
         if (loopInitializer && ts.isVariableStatement(loopInitializer) &&
             loopInitializer.declarationList.declarations.some((declaration) => !declaration.initializer)) return false;
         const bodyPreludeStatements = loopBody.slice(0, -1);
@@ -35031,7 +35030,8 @@ class Emitter {
                 loopInitializerCaptures,
                 loopIncrementor,
             )) return true;
-        if (loopInitializer && ts.isVariableStatement(loopInitializer) && loopBreakSkipsIncrementor &&
+        if (loopInitializer && ts.isVariableStatement(loopInitializer) &&
+            (loopBreakSkipsIncrementor || (loopBodyContinues && loopContinueIncrementorSupported)) &&
             awaitExpressions.length === 3 &&
             this.emitAsyncAwaitLoopConditionConditionalContinue(
                 buf,
@@ -35044,6 +35044,7 @@ class Emitter {
                 ts.isThrowStatement(fallthrough),
                 loopInitializer,
                 loopInitializerCaptures,
+                loopIncrementor,
             )) return true;
         if (loopInitializer && ts.isVariableStatement(loopInitializer) &&
             (loopBreakSkipsIncrementor || (loopBodyContinues && loopContinueIncrementorSupported)) &&